Game screen is rendering properly

+++ source/doom/hu_lib.cpp
@@ -101,7 +101,7 @@ HUlib_drawTextLine
     x = l->x;
     for (i = 0;i < l->len;i++)
     {
-        c = toupper(l->l[i]);
+        c = static_cast<uint8>(toupper(l->l[i]));
         if (c != ' '
             && c >= l->sc
             && c <= '_')
